Fix bugs and change TrianglePoint to float values

+++ src/main/java/ru/vsu/cs/cg/rasterizerdemo/rasterizer/TrianglePoint.java
@@ -3,11 +3,11 @@ package ru.vsu.cs.cg.rasterizerdemo.rasterizer;
 import javafx.scene.paint.Color;
 
 public class TrianglePoint {
-    int x;
-    int y;
+    float x;
+    float y;
     private Color color;
 
-    public TrianglePoint(int x, int y, Color color) {
+    public TrianglePoint(float x, float y, Color color) {
         this.x = x;
         this.y = y;
         this.color = color;
@@ -17,19 +17,19 @@ public class TrianglePoint {
         this(x, y, Color.BLACK);
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
